calories

+++ cals/cals.py
@@ -10,8 +10,8 @@ def get_totals():
     with open("calories.csv", "r") as f:
 
         for row in csv.reader(f):
-            total_cals += (int(row[2]) * int(row[0]))
-            total_protein += (int(row[3]) * int(row[0]))
+            total_cals += (float(row[2]) * float(row[0]))
+            total_protein += (float(row[3]) * float(row[0]))
     
     return total_cals, total_protein
